% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{duplicates}
\alias{duplicates}
\title{Identify duplicates and provide a summary}
\usage{
duplicates(df, ..., output = "summary")
}
\arguments{
\item{df}{Data frame or tibble object.}

\item{output}{Character. Can be either 'summary', 'flag' or 'all'. 'summary' provides a df with the number of duplicates based on the columns provided. 'flag' provides the original df with a new column 'dup_flag' identifying the number of times the row is in the df and 'rank_flag' which indicates the occurrence of the row. Default is summary.}

\item{cols}{Column names.}
}
\value{
A data frame with the summary or the data frame with the flags. If output is 'all' the return is a list with two data frames.
}
\description{
This function can be used to assess the number of duplicates in a dataset based on one or several columns.
}
