% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry_elections.R
\name{lookup_ine_geo_code}
\alias{lookup_ine_geo_code}
\title{Get geographical codes}
\usage{
lookup_ine_geo_code(geo_area)
}
\arguments{
\item{geo_area}{Character. INE geographical code to retrieve the results from a given area.}
}
\value{
A data frames containing the geographical codes.
}
\description{
This function gets INE geographical codes.
}
