% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{plot_colors_party}
\alias{plot_colors_party}
\title{Generate names vector for Spanish parties' colours}
\usage{
plot_colors_party()
}
\value{
A named character vector with spanish party colours.
}
\description{
This function gets election data from a repository.
}
