% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_prepare_election_data.R
\name{read_elec_data}
\alias{read_elec_data}
\title{Read and prepare election data}
\usage{
read_elec_data(path)
}
\arguments{
\item{path}{Path to the DTA file from MIR}
}
\value{
A data frame with the election data
}
\description{
This function gets election data from original DTA files
}
