% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_elections_seats.R
\name{seats_dhont}
\alias{seats_dhont}
\title{Compute D'Hondt rule to assign seats}
\usage{
seats_dhont(data, party, votes, seats, blanks = 0, barrier = 0)
}
\arguments{
\item{data}{Data frame. Input data frame with two columns party and votes.}

\item{party}{Character. Name of the column for parties.}

\item{votes}{Character. Name of the columns with the number of votes.}

\item{seats}{Numeric. Number of seats to be assigned.}

\item{blanks}{Numeric. Number of blank votes in the circunscription. Default is zero.}

\item{barrier}{Numeric. Vote percentafe for electoral barrier. Default is zero.}
}
\value{
A data frame with three columns party, votes and seats.
}
\description{
This function computes the seats in election data.
}
