% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_sql_management.R
\name{sql_insert_batches}
\alias{sql_insert_batches}
\title{Upload df to table per batches}
\usage{
sql_insert_batches(con, table_name, df, batch_size)
}
\arguments{
\item{con}{Connection to db.}

\item{table_name}{Name of the db table to insert the cases.}

\item{df}{Data frame to be uploaded.}

\item{batch_size}{Size of the batch}
}
\value{
Message to indicate the cases uploaded.
}
\description{
It uploads the data to the MariaDB/MySQL using dbx
}
