% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry_elections.R
\name{elections}
\alias{elections}
\title{Get election data}
\usage{
elections(
  elec,
  year,
  month = NULL,
  ine_geo_code = NULL,
  by = NULL,
  output = "summary"
)
}
\arguments{
\item{elec}{Character. Type of election "congreso", "senado", "municipio", "referendum", "euro", "cabildo"}

\item{year}{Numeric. Year of the election}

\item{month}{Numeric. Month of the election only necessary for 2019 "congreso" and "senado". Default is NULL.}

\item{ine_geo_code}{Character. INE geographical code to retrieve the results from a given area.}

\item{by}{Character. Level at which data is retrieved: "nacional", "comunidad", "provincia", "municipio", "distrito", "seccion", "mesa".}

\item{output}{Character. Type of output "summary" or "all".}
}
\value{
A data frames containing the election results.
}
\description{
This function gets election data from a repository.
}
